+++ components/style/color/mod.rs
@@ -6,6 +6,7 @@
 
 /// cbindgen:ignore
 pub mod convert;
+pub mod mix;
 
 use std::fmt::{self, Write};
 use style_traits::{CssWriter, ToCss};
